Add files via upload
this is the full source code + sql database just simple download and extract it and past it in xamp/hdocs folder.
then come back to your browser  type in search bar localhost/sunplaza/.   for adding the data base search localhost/phpmyadmin. create a database with the name of sunplaza and emport the database from the folder which you place in xamp hdocs folder and enjoy.Thank you.